Cambios de vuelos
Se crea controlador de vuelos
Se crea servicio  de vuelos
Se modifican el tipo de id de las entidades de Integer a Long
Se crea metodo findAllByOrden de forma generica y se implementa en los
servicios y controladores relacionados

+++ src/main/java/com/aerolinea/api/model/repository/ClienteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.aerolinea.api.model.entity.Cliente;
 
-public interface ClienteRepository extends CrudRepository<Cliente, Integer> {
+public interface ClienteRepository extends CrudRepository<Cliente, Long> {
 	public List<Cliente> findByNombre(String nombre);
 
 	public List<Cliente> findByapPaterno(String apellido);
