+++ src/main/java/com/aerolinea/api/model/repository/DestinoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.aerolinea.api.model.entity.Destino;
 
-public interface DestinoRepository extends CrudRepository<Destino, Integer> {
+public interface DestinoRepository extends CrudRepository<Destino, Long> {
 	
 	@Query(value = "select * from destinos where lower(translate(ciudad, 'áéíóú', 'aeiou')) like %?1%",nativeQuery = true)
 	public List<Destino> findByCiudadLike(String ciudad);
