+++ src/main/java/com/aerolinea/api/model/repository/EmpleadoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import com.aerolinea.api.model.entity.Empleado;
 import com.aerolinea.api.model.entity.Puesto;
 
-public interface EmpleadoRepository extends CrudRepository<Empleado, Integer>{
+public interface EmpleadoRepository extends CrudRepository<Empleado, Long>{
 	public Integer countByPuesto(Puesto puesto);
 }
