+++ src/main/java/com/aerolinea/api/model/repository/EquipajeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.aerolinea.api.model.entity.Equipaje;;
 
-public interface EquipajeRepository extends CrudRepository<Equipaje, Integer>{
+public interface EquipajeRepository extends CrudRepository<Equipaje, Long>{
 	
 }
