+++ src/main/java/com/aerolinea/api/service/utils/rowmapper/DestinoRowMapper.java
@@ -12,7 +12,7 @@ public class DestinoRowMapper implements RowMapper<Destino>{
 	@Override
 	public Destino mapRow(ResultSet rs, int rowNum) throws SQLException {
 		Destino destino=new Destino();
-		destino.setId(rs.getInt("id"));
+		destino.setId(rs.getLong("id"));
 		destino.setCiudad(rs.getString("ciudad"));
 		destino.setPais(rs.getString("pais"));
 		destino.setDisponible(rs.getBoolean("disponible"));
